split import from server action
