+++ app/api/oauth/[route]/route.ts
@@ -1,4 +1,3 @@
-import { OAuthClientMetadata } from "@atproto/oauth-client-node";
 import { createIdentity } from "actions/createIdentity";
 import { subscribeToPublication } from "app/lish/subscribe";
 import { drizzle } from "drizzle-orm/postgres-js";
@@ -11,30 +10,15 @@ import { setAuthToken } from "src/auth";
 
 import { supabaseServerClient } from "supabase/serverClient";
 import { URLSearchParams } from "url";
+import {
+  ActionAfterSignIn,
+  parseActionFromSearchParam,
+} from "./afterSignInActions";
 
 type OauthRequestClientState = {
   redirect: string | null;
   action: ActionAfterSignIn | null;
 };
-export type ActionAfterSignIn = {
-  action: "subscribe";
-  publication: string;
-};
-
-export function encodeActionToSearchParam(actions: ActionAfterSignIn): string {
-  return encodeURIComponent(JSON.stringify(actions));
-}
-
-function parseActionFromSearchParam(
-  param: string | null,
-): ActionAfterSignIn | null {
-  if (!param) return null;
-  try {
-    return JSON.parse(decodeURIComponent(param)) as ActionAfterSignIn;
-  } catch {
-    return null;
-  }
-}
 
 export async function GET(
   req: NextRequest,
